refactor: new waiter functionality in helmAddonApplier
